// apps/web/src/api/generated/organizer-tournaments/updateTournamentDraftDtoMode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type UpdateTournamentDraftDtoMode = typeof UpdateTournamentDraftDtoMode[keyof typeof UpdateTournamentDraftDtoMode];


export const UpdateTournamentDraftDtoMode = {
  ONLINE: 'ONLINE',
  ONSITE: 'ONSITE',
} as const;