// apps/web/src/api/generated/organizer-tournaments/updateTournamentDraftDtoVisibility.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type UpdateTournamentDraftDtoVisibility = typeof UpdateTournamentDraftDtoVisibility[keyof typeof UpdateTournamentDraftDtoVisibility];


export const UpdateTournamentDraftDtoVisibility = {
  PUBLIC: 'PUBLIC',
  UNLISTED: 'UNLISTED',
  PRIVATE: 'PRIVATE',
} as const;